Finish expressions, flatten directory structure, cleanup.
Expressions no longer get their own directories now; they just get
tacked onto the filename, e.g. 133-grin.png.  This means that
1234-*.png matches all of #1234's sprites, and 1234-abc-*.png matches
all of #1234's sprites in abc form; the problem with this is that if we
ever get a Pokémon with expressions for its alternate forms and no name
for its default form, we won't be able to separate them.  (This is
already the case with 40 vs 40-mama and 352 vs 352-purple, actually, but
as those are both nonstandard forms, this behaviour is welcomed for
veekun's purposes.  At least until the next wave of PMD games comes out
and breaks this system.)