rip.sh: rip from both Sky and Blue, then merge.
If you actually do this, you'll discover that there are a surprising
amount of sprites that differ between Blue and Sky, including a bunch of
standard-expression sprites — and I haven't even found most of those!
For the most part, the differences are minor — palette tweaks, nostrils
getting shifted by a couple pixels, etc — but it still makes this more
difficult.  I guess I should expand this to rip /all/ the Blue sprites?
And, if there are too many unique ones, don't bother trying to merge
them with the Sky sprites?  Agh.